requirements.txt
PyQt5==5.15.11
lxml==5.3.0
openpyxl==3.1.5
python-docx==1.1.2